cmd/kubectl-gadget: Move consts to utils package
