+++ cmd/kubectl-gadget/utils/consts.go
@@ -12,8 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package main
+package utils
 
 const (
-	gadgetNamespace string = "gadget"
+	GadgetNamespace string = "gadget"
 )
